Completed Activity 2, tested standard threading code corectly, all tests pass.
