+++ src/CMakeLists.txt
@@ -1,6 +1,7 @@
 # This is the main binary. List your C files here.
 add_executable(hello_freertos
     main.c
+    threading.c
     )
 
 pico_set_program_name(hello_freertos "test")
