+++ test/unity_config.h
@@ -19,6 +19,7 @@ void unityOutputChar(char);
 void unityOutputFlush();
 void unityOutputComplete();
 
+// NOTE: Commented out due to multipe definition error. Seems to be working wtihout these.
 // #define UNITY_OUTPUT_START()    unityOutputStart()
 // #define UNITY_OUTPUT_CHAR(c)    unityOutputChar(c)
 // #define UNITY_OUTPUT_FLUSH()    unityOutputFlush()
